Backend: finalize normalized student/course enrollment flow and testing

+++ database/init/01_schema.sql
@@ -17,7 +17,7 @@ CREATE TABLE Courses (
 CREATE TABLE Enrollments (
     student_id INT,
     course_id INT,
-    score INT NOT NULL,
+    score DECIMAL(5,2) NOT NULL,
     PRIMARY KEY (student_id, course_id),
     FOREIGN KEY (student_id) REFERENCES Students(student_id),
     FOREIGN KEY (course_id) REFERENCES Courses(course_id),
